Rename utils to process
